[WIP] Webpack 4, react-error-overlay, react-loadable (#4639)
Webpack 4, react-error-overlay, react-loadable (major)

+++ test/integration/basic/test/error-recovery.js
@@ -1,101 +1,121 @@
 /* global describe, it, expect */
 import webdriver from 'next-webdriver'
 import { join } from 'path'
-import { check, File, waitFor } from 'next-test-utils'
+import { check, File, waitFor, getReactErrorOverlayContent } from 'next-test-utils'
 
 export default (context, render) => {
   describe('Error Recovery', () => {
     it('should detect syntax errors and recover', async () => {
-      const browser = await webdriver(context.appPort, '/hmr/about')
-      const text = await browser
-        .elementByCss('p').text()
-      expect(text).toBe('This is the about page.')
-
+      let browser
       const aboutPage = new File(join(__dirname, '../', 'pages', 'hmr', 'about.js'))
-      aboutPage.replace('</div>', 'div')
+      try {
+        browser = await webdriver(context.appPort, '/hmr/about')
+        const text = await browser
+          .elementByCss('p').text()
+        expect(text).toBe('This is the about page.')
 
-      await check(
-        () => browser.elementByCss('body').text(),
-        /Unterminated JSX contents/
-      )
+        aboutPage.replace('</div>', 'div')
 
-      aboutPage.restore()
+        await waitFor(3000)
 
-      await check(
-        () => browser.elementByCss('body').text(),
-        /This is the about page/
-      )
+        expect(await getReactErrorOverlayContent(browser)).toMatch(/Unterminated JSX contents/)
 
-      browser.close()
+        aboutPage.restore()
+
+        await check(
+          () => browser.elementByCss('body').text(),
+          /This is the about page/
+        )
+      } finally {
+        aboutPage.restore()
+
+        if (browser) {
+          browser.close()
+        }
+      }
     })
 
     it('should not show the default HMR error overlay', async () => {
-      const browser = await webdriver(context.appPort, '/hmr/about')
-      const text = await browser
-        .elementByCss('p').text()
-      expect(text).toBe('This is the about page.')
-
+      let browser
       const aboutPage = new File(join(__dirname, '../', 'pages', 'hmr', 'about.js'))
-      aboutPage.replace('</div>', 'div')
+      try {
+        browser = await webdriver(context.appPort, '/hmr/about')
+        const text = await browser
+          .elementByCss('p').text()
+        expect(text).toBe('This is the about page.')
 
-      await check(
-        () => browser.elementByCss('body').text(),
-        /Unterminated JSX contents/
-      )
+        aboutPage.replace('</div>', 'div')
 
-      await waitFor(2000)
+        await waitFor(3000)
 
-      // Check for the error overlay
-      const bodyHtml = await browser.elementByCss('body').getAttribute('innerHTML')
-      expect(bodyHtml.includes('webpack-hot-middleware-clientOverlay')).toBeFalsy()
+        expect(await getReactErrorOverlayContent(browser)).toMatch(/Unterminated JSX contents/)
 
-      aboutPage.restore()
-      browser.close()
+        await waitFor(2000)
+
+        // Check for the error overlay
+        const bodyHtml = await browser.elementByCss('body').getAttribute('innerHTML')
+        expect(bodyHtml.includes('webpack-hot-middleware-clientOverlay')).toBeFalsy()
+      } finally {
+        aboutPage.restore()
+        if (browser) {
+          browser.close()
+        }
+      }
     })
 
     it('should show the error on all pages', async () => {
       const aboutPage = new File(join(__dirname, '../', 'pages', 'hmr', 'about.js'))
-      aboutPage.replace('</div>', 'div')
+      let browser
+      try {
+        aboutPage.replace('</div>', 'div')
 
-      const browser = await webdriver(context.appPort, '/hmr/contact')
+        browser = await webdriver(context.appPort, '/hmr/contact')
 
-      await check(
-        () => browser.elementByCss('body').text(),
-        /Unterminated JSX contents/
-      )
+        await waitFor(3000)
 
-      aboutPage.restore()
+        expect(await getReactErrorOverlayContent(browser)).toMatch(/Unterminated JSX contents/)
 
-      await check(
-        () => browser.elementByCss('body').text(),
-        /This is the contact page/
-      )
+        aboutPage.restore()
 
-      browser.close()
+        await check(
+          () => browser.elementByCss('body').text(),
+          /This is the contact page/
+        )
+      } finally {
+        aboutPage.restore()
+        if (browser) {
+          browser.close()
+        }
+      }
     })
 
     it('should detect runtime errors on the module scope', async () => {
-      const browser = await webdriver(context.appPort, '/hmr/about')
-      const text = await browser
-        .elementByCss('p').text()
-      expect(text).toBe('This is the about page.')
-
+      let browser
       const aboutPage = new File(join(__dirname, '../', 'pages', 'hmr', 'about.js'))
-      aboutPage.replace('export', 'aa=20;\nexport')
+      try {
+        browser = await webdriver(context.appPort, '/hmr/about')
+        const text = await browser
+          .elementByCss('p').text()
+        expect(text).toBe('This is the about page.')
 
-      await check(
-        () => browser.elementByCss('body').text(),
-        /aa is not defined/
-      )
+        aboutPage.replace('export', 'aa=20;\nexport')
 
-      aboutPage.restore()
+        await waitFor(3000)
 
-      await check(
-        () => browser.elementByCss('body').text(),
-        /This is the about page/
-      )
+        expect(await getReactErrorOverlayContent(browser)).toMatch(/aa is not defined/)
 
-      browser.close()
+        aboutPage.restore()
+
+        await check(
+          () => browser.elementByCss('body').text(),
+          /This is the about page/
+        )
+      } finally {
+        aboutPage.restore()
+        if (browser) {
+          browser.close()
+        }
+      }
     })
 
     it('should recover from errors in the render function', async () => {
@@ -107,10 +127,9 @@ export default (context, render) => {
       const aboutPage = new File(join(__dirname, '../', 'pages', 'hmr', 'about.js'))
       aboutPage.replace('return', 'throw new Error("an-expected-error");\nreturn')
 
-      await check(
-        () => browser.elementByCss('body').text(),
-        /an-expected-error/
-      )
+      await waitFor(3000)
+
+      expect(await getReactErrorOverlayContent(browser)).toMatch(/an-expected-error/)
 
       aboutPage.restore()
 
@@ -131,10 +150,9 @@ export default (context, render) => {
       const aboutPage = new File(join(__dirname, '../', 'pages', 'hmr', 'about.js'))
       aboutPage.replace('export default', 'export default "not-a-page"\nexport const fn = ')
 
-      await check(
-        () => browser.elementByCss('body').text(),
-        /The default export is not a React Component/
-      )
+      await waitFor(3000)
+
+      expect(await browser.elementByCss('body').text()).toMatch(/The default export is not a React Component/)
 
       aboutPage.restore()
 
@@ -174,10 +192,9 @@ export default (context, render) => {
       const browser = await webdriver(context.appPort, '/hmr')
       await browser.elementByCss('#error-in-gip-link').click()
 
-      await check(
-        () => browser.elementByCss('body').text(),
-        /an-expected-error-in-gip/
-      )
+      await waitFor(1500)
+
+      expect(await getReactErrorOverlayContent(browser)).toMatch(/an-expected-error-in-gip/)
 
       const erroredPage = new File(join(__dirname, '../', 'pages', 'hmr', 'error-in-gip.js'))
       erroredPage.replace('throw error', 'return {}')
@@ -194,10 +211,9 @@ export default (context, render) => {
     it('should recover after an error reported via SSR', async () => {
       const browser = await webdriver(context.appPort, '/hmr/error-in-gip')
 
-      await check(
-        () => browser.elementByCss('body').text(),
-        /an-expected-error-in-gip/
-      )
+      await waitFor(1500)
+
+      expect(await getReactErrorOverlayContent(browser)).toMatch(/an-expected-error-in-gip/)
 
       const erroredPage = new File(join(__dirname, '../', 'pages', 'hmr', 'error-in-gip.js'))
       erroredPage.replace('throw error', 'return {}')
@@ -212,24 +228,37 @@ export default (context, render) => {
     })
 
     it('should recover from 404 after a page has been added', async () => {
-      const browser = await webdriver(context.appPort, '/hmr/new-page')
+      let browser
+      let newPage
+      try {
+        browser = await webdriver(context.appPort, '/hmr/new-page')
 
-      await check(
-        () => browser.elementByCss('body').text(),
-        /This page could not be found/
-      )
+        expect(await browser.elementByCss('body').text()).toMatch(/This page could not be found/)
 
-      // Add the page
-      const newPage = new File(join(__dirname, '../', 'pages', 'hmr', 'new-page.js'))
-      newPage.write('export default () => (<div>the-new-page</div>)')
+        // Add the page
+        newPage = new File(join(__dirname, '../', 'pages', 'hmr', 'new-page.js'))
+        newPage.write('export default () => (<div id="new-page">the-new-page</div>)')
 
-      await check(
-        () => browser.elementByCss('body').text(),
-        /the-new-page/
-      )
+        await check(
+          () => {
+            if (!browser.hasElementById('new-page')) {
+              throw new Error('waiting')
+            }
 
-      newPage.delete()
-      browser.close()
+            return browser.elementByCss('body').text()
+          },
+          /the-new-page/
+        )
+
+        // expect(await browser.elementByCss('body').text()).toMatch(/the-new-page/)
+      } finally {
+        if (newPage) {
+          newPage.delete()
+        }
+        if (browser) {
+          browser.close()
+        }
+      }
     })
   })
 }
